Complete Exercise 3
Use `app.param()` to find the lion with the matching `id` parameter,
then attach the matching lion to the request object and call `next()`.
Then in the `GET /lion/:id` route, send back `req.lion`.

Create a middleware function to catch and handle errors, and register it
as the last middleware function on `app.`