Complete Exercise 11
Build out the following functions in `auth/auth.js`:

getFreshUser()

You will have access to `req.user` in this function because we will use
decodeToken befoer this function in the middleware stack. `req.user`
will be an object that contains the user’s ID. We want the full user
object. If no user is found, it was a valid JWT but didn’t decode to a
real user in our database. Either the user was deleted since the client
got the JWT, or it was a JWT from some other source. Find the user in
the database, and if found update `req.user` with the user object
returned from the database. Otherwise, return the appropriate status
code.

verifyUser()

If no username or password was provided, stop/do not proceed further.
Otherwise, query the database for the user. If the user is not found,
return the approriate status code and message. If the incorrect password
was provided, return the approriate status code and message. Otherwise,
set `req.user` to the returned user object.

Build out the following function in `auth/controller.js`

signin()

Create the JWT token and return it to the client.